+++ spec/models/gadget_spec.rb
@@ -23,43 +23,4 @@ RSpec.describe Gadget, type: :model do
       }.to(change { gadget.versions.size }.by(1))
     end
   end
-
-  describe "#changed_notably?", versioning: true do
-    context "new record" do
-      it "returns true" do
-        g = Gadget.new(created_at: Time.now)
-        expect(g.paper_trail.changed_notably?).to eq(true)
-      end
-    end
-
-    context "persisted record without update timestamps" do
-      it "only acknowledges non-ignored attrs" do
-        gadget = Gadget.create!(created_at: Time.now)
-        gadget.name = "Wrench"
-        expect(gadget.paper_trail.changed_notably?).to be true
-      end
-
-      it "does not acknowledge ignored attr (brand)" do
-        gadget = Gadget.create!(created_at: Time.now)
-        gadget.brand = "Acme"
-        expect(gadget.paper_trail.changed_notably?).to be false
-      end
-    end
-
-    context "persisted record with update timestamps" do
-      it "only acknowledges non-ignored attrs" do
-        gadget = Gadget.create!(created_at: Time.now)
-        gadget.name = "Wrench"
-        gadget.updated_at = Time.now
-        expect(gadget.paper_trail.changed_notably?).to be true
-      end
-
-      it "does not acknowledge ignored attrs and timestamps only" do
-        gadget = Gadget.create!(created_at: Time.now)
-        gadget.brand = "Acme"
-        gadget.updated_at = Time.now
-        expect(gadget.paper_trail.changed_notably?).to be false
-      end
-    end
-  end
 end
